Groups with Roles feature (#1859)
Groups with Roles feature

Introduce ability to add roles to groups. Group members inherit the permissions attached to the group's roles.

Notable changes:
- Several new Roles for Work, Collection, and User management
- Role permission integration with Hyrax permission logic (Sipity, access controls, CanCan, etc.)
- Hyku::Group replaced with already existing Hyrax::Group
- Easily assign roles to individual users
- Extended UI
- Both pre-configured and customizable groups
- Scripts to migrate existing Hyku applications
- More

Documentation for this feature can be found in the Hyku GitHub Wiki here:

https://github.com/samvera/hyku/wiki/Groups-with-Roles-Feature

This work is a joint contribution of PALNI and PALCI's Hyku for Consortia project.

+++ app/controllers/admin/groups_controller.rb
@@ -1,27 +1,37 @@
 # frozen_string_literal: true
 
 module Admin
-  class GroupsController < AdminController
-    before_action :load_group, only: %i[edit update remove destroy]
+  class GroupsController < ApplicationController
+    load_and_authorize_resource(
+      class: '::Hyrax::Group',
+      instance_name: :group,
+      only: %i[create edit update remove destroy]
+    )
+    layout 'hyrax/dashboard'
 
     def index
+      authorize! :read, Hyrax::Group
       add_breadcrumb t(:'hyrax.controls.home'), root_path
       add_breadcrumb t(:'hyrax.dashboard.breadcrumbs.admin'), hyrax.dashboard_path
       add_breadcrumb t(:'hyku.admin.groups.title.index'), admin_groups_path
-      @groups = Hyku::Group.search(params[:q]).page(page_number).per(page_size)
+      @groups = Hyrax::Group.search(params[:q]).page(page_number).per(page_size)
     end
 
     def new
       add_breadcrumb t(:'hyrax.controls.home'), root_path
       add_breadcrumb t(:'hyrax.dashboard.breadcrumbs.admin'), hyrax.dashboard_path
       add_breadcrumb t(:'hyku.admin.groups.title.new'), new_admin_group_path
-      @group = Hyku::Group.new
+      @group = Hyrax::Group.new
     end
 
     def create
-      new_group = Hyku::Group.new(group_params)
+      new_group = Hyrax::Group.new(group_params)
+      new_group.name = group_params[:humanized_name].tr(" ", "_").downcase
       if new_group.save
-        redirect_to admin_groups_path, notice: t('hyku.admin.groups.flash.create.success', group: new_group.name)
+        redirect_to(
+          admin_groups_path,
+          notice: t('hyku.admin.groups.flash.create.success', group: new_group.humanized_name)
+        )
       elsif new_group.invalid?
         redirect_to new_admin_group_path, alert: t('hyku.admin.groups.flash.create.invalid')
       else
@@ -36,11 +46,12 @@ module Admin
     end
 
     def update
+      @group.name = group_params[:humanized_name].tr(" ", "_").downcase
       if @group.update(group_params)
-        redirect_to admin_groups_path, notice: t('hyku.admin.groups.flash.update.success', group: @group.name)
+        redirect_to admin_groups_path, notice: t('hyku.admin.groups.flash.update.success', group: @group.humanized_name)
       else
         redirect_to edit_admin_group_path(@group), flash: {
-          error: t('hyku.admin.groups.flash.update.failure', group: @group.name)
+          error: t('hyku.admin.groups.flash.update.failure', group: @group.humanized_name)
         }
       end
     end
@@ -50,25 +61,30 @@ module Admin
       add_breadcrumb t(:'hyrax.dashboard.breadcrumbs.admin'), hyrax.dashboard_path
       add_breadcrumb t(:'hyku.admin.groups.title.edit'), edit_admin_group_path
       add_breadcrumb t(:'hyku.admin.groups.title.remove'), request.path
+
+      flash.now[:alert] = t('hyku.admin.groups.flash.remove.cannot_destroy_default_group') if @group.default_group?
     end
 
     def destroy
+      return redirect_back(fallback_location: admin_groups_path) if @group.default_group?
       if @group.destroy
-        redirect_to admin_groups_path, notice: t('hyku.admin.groups.flash.destroy.success', group: @group.name)
+        redirect_to(
+          admin_groups_path,
+          notice: t('hyku.admin.groups.flash.destroy.success', group: @group.humanized_name)
+        )
       else
-        logger.error("Hyku::Group id:#{@group.id} could not be destroyed")
-        redirect_to admin_groups_path flash: { error: t('hyku.admin.groups.flash.destroy.failure', group: @group.name) }
+        logger.error("Hyrax::Group id:#{@group.id} could not be destroyed")
+        redirect_to(
+          admin_groups_path,
+          flash: { error: t('hyku.admin.groups.flash.destroy.failure', group: @group.humanized_name) }
+        )
       end
     end
 
     private
 
-      def load_group
-        @group = Hyku::Group.find_by(id: params[:id])
-      end
-
       def group_params
-        params.require(:hyku_group).permit(:name, :description)
+        params.require(:group).permit(:name, :humanized_name, :description)
       end
 
       def page_number
